[AMQ-8354] Extract replication logic to a separate class.
