+++ activemq-broker/src/main/java/org/apache/activemq/replica/source/ReplicaSourceBroker.java
@@ -21,12 +21,9 @@ import org.apache.activemq.broker.ConnectionContext;
 import org.apache.activemq.broker.ConsumerBrokerExchange;
 import org.apache.activemq.broker.ProducerBrokerExchange;
 import org.apache.activemq.broker.region.Destination;
-import org.apache.activemq.broker.region.DurableTopicSubscription;
 import org.apache.activemq.broker.region.MessageReference;
 import org.apache.activemq.broker.region.PrefetchSubscription;
-import org.apache.activemq.broker.region.QueueBrowserSubscription;
 import org.apache.activemq.broker.region.Subscription;
-import org.apache.activemq.broker.region.Topic;
 import org.apache.activemq.command.ActiveMQDestination;
 import org.apache.activemq.command.ActiveMQMessage;
 import org.apache.activemq.command.ConnectionId;
@@ -37,57 +34,44 @@ import org.apache.activemq.command.MessageAck;
 import org.apache.activemq.command.MessageId;
 import org.apache.activemq.command.RemoveSubscriptionInfo;
 import org.apache.activemq.command.TransactionId;
-import org.apache.activemq.filter.DestinationMap;
-import org.apache.activemq.filter.DestinationMapEntry;
-import org.apache.activemq.replica.util.DestinationExtractor;
 import org.apache.activemq.replica.MutativeRoleBroker;
-import org.apache.activemq.replica.util.ReplicaEvent;
-import org.apache.activemq.replica.util.ReplicaEventSerializer;
-import org.apache.activemq.replica.util.ReplicaEventType;
 import org.apache.activemq.replica.ReplicaPolicy;
 import org.apache.activemq.replica.ReplicaReplicationDestinationSupplier;
-import org.apache.activemq.replica.util.ReplicaRole;
 import org.apache.activemq.replica.ReplicaRoleManagement;
+import org.apache.activemq.replica.util.ReplicaEventType;
+import org.apache.activemq.replica.util.ReplicaRole;
 import org.apache.activemq.replica.util.ReplicaSupport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Collectors;
 
 import static org.apache.activemq.replica.util.ReplicaSupport.MAIN_REPLICATION_QUEUE_NAME;
 
 public class ReplicaSourceBroker extends MutativeRoleBroker {
-    private static final DestinationMapEntry<Boolean> IS_REPLICATED = new DestinationMapEntry<>() {
-    }; // used in destination map to indicate mirrored status
 
     private static final Logger logger = LoggerFactory.getLogger(ReplicaSourceBroker.class);
     private final ScheduledExecutorService heartBeatPoller = Executors.newSingleThreadScheduledExecutor();
-    private final ReplicaEventSerializer eventSerializer = new ReplicaEventSerializer();
-    private final AtomicBoolean initialized = new AtomicBoolean();
 
-    private final ReplicationMessageProducer replicationMessageProducer;
+    private final ReplicaEventReplicator replicaEventReplicator;
     private final ReplicaSequencer replicaSequencer;
     private final ReplicaReplicationDestinationSupplier destinationSupplier;
     private final ReplicaPolicy replicaPolicy;
     private final ReplicaAckHelper replicaAckHelper;
     private ScheduledFuture<?> heartBeatScheduledFuture;
 
-    final DestinationMap destinationsToReplicate = new DestinationMap();
-
-    public ReplicaSourceBroker(Broker broker, ReplicaRoleManagement management, ReplicationMessageProducer replicationMessageProducer,
+    public ReplicaSourceBroker(Broker broker, ReplicaRoleManagement management, ReplicaEventReplicator replicaEventReplicator,
             ReplicaSequencer replicaSequencer, ReplicaReplicationDestinationSupplier destinationSupplier,
             ReplicaPolicy replicaPolicy) {
         super(broker, management);
-        this.replicationMessageProducer = replicationMessageProducer;
+        this.replicaEventReplicator = replicaEventReplicator;
         this.replicaSequencer = replicaSequencer;
         this.destinationSupplier = destinationSupplier;
         this.replicaPolicy = replicaPolicy;
@@ -99,10 +83,10 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
         logger.info("Starting Source broker. " + (role == ReplicaRole.await_ack ? " Awaiting ack." : ""));
 
         initQueueProvider();
-        initialized.compareAndSet(false, true);
+        replicaEventReplicator.initialize();
         replicaSequencer.initialize();
         initializeHeartBeatSender();
-        ensureDestinationsAreReplicated();
+        replicaEventReplicator.ensureDestinationsAreReplicated();
     }
 
     @Override
@@ -116,7 +100,7 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
     public void stop() throws Exception {
         replicaSequencer.deinitialize();
         super.stop();
-        initialized.compareAndSet(true, false);
+        replicaEventReplicator.deinitialize();
     }
 
     @Override
@@ -136,7 +120,7 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
         startAllConnections();
 
         initQueueProvider();
-        initialized.compareAndSet(false, true);
+        replicaEventReplicator.initialize();
         replicaSequencer.initialize();
         initializeHeartBeatSender();
         replicaSequencer.updateMainQueueConsumerStatus();
@@ -144,18 +128,8 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
 
     private void initializeHeartBeatSender() {
         if (replicaPolicy.getHeartBeatPeriod() > 0) {
-            heartBeatScheduledFuture = heartBeatPoller.scheduleAtFixedRate(() -> {
-                try {
-                    enqueueReplicaEvent(
-                            getAdminConnectionContext(),
-                            new ReplicaEvent()
-                                    .setEventType(ReplicaEventType.HEART_BEAT)
-                                    .setEventData(eventSerializer.serializeReplicationData(null))
-                    );
-                } catch (Exception e) {
-                    logger.error("Failed to send heart beat message", e);
-                }
-            }, replicaPolicy.getHeartBeatPeriod(), replicaPolicy.getHeartBeatPeriod(), TimeUnit.MILLISECONDS);
+            heartBeatScheduledFuture = heartBeatPoller.scheduleAtFixedRate(replicaEventReplicator::enqueueHeartBeatEvent,
+                    replicaPolicy.getHeartBeatPeriod(), replicaPolicy.getHeartBeatPeriod(), TimeUnit.MILLISECONDS);
         }
     }
 
@@ -180,203 +154,6 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
         destinationSupplier.initializeSequenceQueue();
     }
 
-    private void ensureDestinationsAreReplicated() throws Exception {
-        for (ActiveMQDestination d : getDurableDestinations()) {
-            if (!shouldReplicateDestination(d)) {
-                continue;
-            }
-            replicateDestinationCreation(getAdminConnectionContext(), d);
-
-            if (!d.isTopic()) {
-                continue;
-            }
-
-            List<DurableTopicSubscription> durableTopicSubscriptions = getDestinations(d).stream().findFirst()
-                    .map(DestinationExtractor::extractTopic)
-                    .map(Topic::getDurableTopicSubs)
-                    .stream()
-                    .map(Map::values)
-                    .flatMap(Collection::stream)
-                    .collect(Collectors.toList());
-            for (DurableTopicSubscription sub : durableTopicSubscriptions) {
-                replicateAddConsumer(getAdminConnectionContext(), sub.getConsumerInfo(), sub.getContext().getClientId());
-            }
-        }
-    }
-
-    private void replicateDestinationCreation(ConnectionContext context, ActiveMQDestination destination) throws Exception {
-        if (destinationsToReplicate.chooseValue(destination) != null) {
-            return;
-        }
-
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.DESTINATION_UPSERT)
-                            .setEventData(eventSerializer.serializeReplicationData(destination))
-            );
-            destinationsToReplicate.put(destination, IS_REPLICATED);
-        } catch (Exception e) {
-            logger.error("Failed to replicate creation of destination {}", destination.getPhysicalName(), e);
-            throw e;
-        }
-    }
-
-    private boolean shouldReplicateDestination(ActiveMQDestination destination) {
-        boolean isReplicationQueue = ReplicaSupport.isReplicationDestination(destination);
-        boolean isAdvisoryDestination = ReplicaSupport.isAdvisoryDestination(destination);
-        boolean isTemporaryDestination = destination.isTemporary();
-        boolean shouldReplicate = !isReplicationQueue && !isAdvisoryDestination && !isTemporaryDestination;
-        String reason = shouldReplicate ? "" : " because ";
-        if (isReplicationQueue) reason += "it is a replication queue";
-        if (isAdvisoryDestination) reason += "it is an advisory destination";
-        if (isTemporaryDestination) reason += "it is a temporary destination";
-        logger.debug("Will {}replicate destination {}{}", shouldReplicate ? "" : "not ", destination, reason);
-        return shouldReplicate;
-    }
-
-    private boolean isReplicatedDestination(ActiveMQDestination destination) {
-        if (destinationsToReplicate.chooseValue(destination) == null) {
-            logger.debug("{} is not a replicated destination", destination.getPhysicalName());
-            return false;
-        }
-        return true;
-    }
-
-    public void replicateSend(ConnectionContext context, Message message, TransactionId transactionId) throws Exception {
-        try {
-            TransactionId originalTransactionId = message.getTransactionId();
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.MESSAGE_SEND)
-                            .setEventData(eventSerializer.serializeMessageData(message))
-                            .setTransactionId(transactionId)
-                            .setReplicationProperty(ReplicaSupport.MESSAGE_ID_PROPERTY, message.getMessageId().toProducerKey())
-                            .setReplicationProperty(ReplicaSupport.IS_ORIGINAL_MESSAGE_SENT_TO_QUEUE_PROPERTY,
-                                    message.getDestination().isQueue())
-                            .setReplicationProperty(ReplicaSupport.ORIGINAL_MESSAGE_DESTINATION_PROPERTY,
-                                    message.getDestination().toString())
-                            .setReplicationProperty(ReplicaSupport.IS_ORIGINAL_MESSAGE_IN_XA_TRANSACTION_PROPERTY,
-                                    originalTransactionId != null && originalTransactionId.isXATransaction())
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate message {} for destination {}", message.getMessageId(), message.getDestination().getPhysicalName(), e);
-            throw e;
-        }
-    }
-
-    public boolean needToReplicateSend(ConnectionContext connectionContext, Message message) {
-        if (isReplicaContext(connectionContext)) {
-            return false;
-        }
-        if (ReplicaSupport.isReplicationDestination(message.getDestination())) {
-            return false;
-        }
-        if (message.getDestination().isTemporary()) {
-            return false;
-        }
-        if (message.isAdvisory()) {
-            return false;
-        }
-        if (!message.isPersistent()) {
-            return false;
-        }
-
-        return true;
-    }
-
-    private void replicateBeginTransaction(ConnectionContext context, TransactionId xid) throws Exception {
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.TRANSACTION_BEGIN)
-                            .setEventData(eventSerializer.serializeReplicationData(xid))
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate begin of transaction [{}]", xid);
-            throw e;
-        }
-    }
-
-    private void replicatePrepareTransaction(ConnectionContext context, TransactionId xid) throws Exception {
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.TRANSACTION_PREPARE)
-                            .setEventData(eventSerializer.serializeReplicationData(xid))
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate transaction prepare [{}]", xid);
-            throw e;
-        }
-    }
-
-    private void replicateForgetTransaction(ConnectionContext context, TransactionId xid) throws Exception {
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.TRANSACTION_FORGET)
-                            .setEventData(eventSerializer.serializeReplicationData(xid))
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate transaction forget [{}]", xid);
-            throw e;
-        }
-    }
-
-    private void replicateRollbackTransaction(ConnectionContext context, TransactionId xid) throws Exception {
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.TRANSACTION_ROLLBACK)
-                            .setEventData(eventSerializer.serializeReplicationData(xid))
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate transaction rollback [{}]", xid);
-            throw e;
-        }
-    }
-
-    private void replicateCommitTransaction(ConnectionContext context, TransactionId xid, boolean onePhase) throws Exception {
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.TRANSACTION_COMMIT)
-                            .setEventData(eventSerializer.serializeReplicationData(xid))
-                            .setReplicationProperty(ReplicaSupport.TRANSACTION_ONE_PHASE_PROPERTY, onePhase)
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate commit of transaction [{}]", xid);
-            throw e;
-        }
-    }
-
-
-    private void replicateDestinationRemoval(ConnectionContext context, ActiveMQDestination destination) throws Exception {
-        if (!isReplicatedDestination(destination)) {
-            return;
-        }
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.DESTINATION_DELETE)
-                            .setEventData(eventSerializer.serializeReplicationData(destination))
-            );
-            destinationsToReplicate.remove(destination, IS_REPLICATED);
-        } catch (Exception e) {
-            logger.error("Failed to replicate remove of destination {}", destination.getPhysicalName(), e);
-            throw e;
-        }
-    }
-
     private void sendFailOverMessage() throws Exception {
         ConnectionContext connectionContext = createConnectionContext();
 
@@ -386,14 +163,7 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
 
         super.beginTransaction(connectionContext, tid);
         try {
-            enqueueReplicaEvent(
-                    connectionContext,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.FAIL_OVER)
-                            .setTransactionId(tid)
-                            .setEventData(eventSerializer.serializeReplicationData(null))
-            );
-
+            replicaEventReplicator.enqueueFailOverEvent(connectionContext, tid);
             updateBrokerRole(connectionContext, tid, ReplicaRole.await_ack);
             super.commitTransaction(connectionContext, tid, true);
         } catch (Exception e) {
@@ -406,7 +176,7 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
     @Override
     public Subscription addConsumer(ConnectionContext context, ConsumerInfo consumerInfo) throws Exception {
         Subscription subscription = super.addConsumer(context, consumerInfo);
-        replicateAddConsumer(context, consumerInfo, context.getClientId());
+        replicaEventReplicator.replicateAddConsumer(context, consumerInfo, context.getClientId());
 
         if (ReplicaSupport.isMainReplicationQueue(consumerInfo.getDestination())) {
             replicaSequencer.updateMainQueueConsumerStatus();
@@ -415,116 +185,44 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
         return subscription;
     }
 
-    private void replicateAddConsumer(ConnectionContext context, ConsumerInfo consumerInfo, String clientId) throws Exception {
-        if (!needToReplicateConsumer(consumerInfo)) {
-            return;
-        }
-        if (ReplicaSupport.isReplicationTransport(context.getConnector())) {
-            return;
-        }
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.ADD_DURABLE_CONSUMER)
-                            .setEventData(eventSerializer.serializeReplicationData(consumerInfo))
-                            .setReplicationProperty(ReplicaSupport.CLIENT_ID_PROPERTY, clientId)
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate adding {}", consumerInfo, e);
-            throw e;
-        }
-    }
-
-    private boolean needToReplicateConsumer(ConsumerInfo consumerInfo) {
-        return consumerInfo.getDestination().isTopic() &&
-                consumerInfo.isDurable() &&
-                !consumerInfo.isNetworkSubscription();
-    }
-
     @Override
     public void removeConsumer(ConnectionContext context, ConsumerInfo consumerInfo) throws Exception {
         super.removeConsumer(context, consumerInfo);
-        replicateRemoveConsumer(context, consumerInfo);
+        replicaEventReplicator.replicateRemoveConsumer(context, consumerInfo);
         if (ReplicaSupport.isMainReplicationQueue(consumerInfo.getDestination())) {
             replicaSequencer.updateMainQueueConsumerStatus();
-        }
-    }
-
-    private void replicateRemoveConsumer(ConnectionContext context, ConsumerInfo consumerInfo) throws Exception {
-        if (!needToReplicateConsumer(consumerInfo)) {
-            return;
-        }
-        if (ReplicaSupport.isReplicationTransport(context.getConnector())) {
-            return;
-        }
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.REMOVE_DURABLE_CONSUMER)
-                            .setEventData(eventSerializer.serializeReplicationData(consumerInfo))
-                            .setReplicationProperty(ReplicaSupport.CLIENT_ID_PROPERTY, context.getClientId())
-                            .setVersion(2)
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate adding {}", consumerInfo, e);
-            throw e;
         }
     }
 
     @Override
     public void removeSubscription(ConnectionContext context, RemoveSubscriptionInfo subscriptionInfo) throws Exception {
         super.removeSubscription(context, subscriptionInfo);
-        replicateRemoveSubscription(context, subscriptionInfo);
-    }
-
-    private void replicateRemoveSubscription(ConnectionContext context, RemoveSubscriptionInfo subscriptionInfo) throws Exception {
-        if (ReplicaSupport.isReplicationTransport(context.getConnector())) {
-            return;
-        }
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.REMOVE_DURABLE_CONSUMER_SUBSCRIPTION)
-                            .setEventData(eventSerializer.serializeReplicationData(subscriptionInfo))
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate removing subscription {}", subscriptionInfo, e);
-            throw e;
-        }
+        replicaEventReplicator.replicateRemoveSubscription(context, subscriptionInfo);
     }
 
     @Override
     public void commitTransaction(ConnectionContext context, TransactionId xid, boolean onePhase) throws Exception {
         super.commitTransaction(context, xid, onePhase);
-        if (xid.isXATransaction()) {
-            replicateCommitTransaction(context, xid, onePhase);
-        }
+        replicaEventReplicator.replicateCommitTransaction(context, xid, onePhase);
     }
 
     @Override
     public int prepareTransaction(ConnectionContext context, TransactionId xid) throws Exception {
         int id = super.prepareTransaction(context, xid);
-        if (xid.isXATransaction()) {
-            replicatePrepareTransaction(context, xid);
-        }
+        replicaEventReplicator.replicatePrepareTransaction(context, xid);
         return id;
     }
 
     @Override
     public void rollbackTransaction(ConnectionContext context, TransactionId xid) throws Exception {
         super.rollbackTransaction(context, xid);
-        if (xid.isXATransaction()) {
-            replicateRollbackTransaction(context, xid);
-        }
+        replicaEventReplicator.replicateRollbackTransaction(context, xid);
     }
 
     @Override
     public void send(ProducerBrokerExchange producerExchange, Message messageSend) throws Exception {
         final ConnectionContext connectionContext = producerExchange.getConnectionContext();
-        if (!needToReplicateSend(connectionContext, messageSend)) {
+        if (!replicaEventReplicator.needToReplicateSend(connectionContext, messageSend)) {
             super.send(producerExchange, messageSend);
             return;
         }
@@ -559,50 +257,27 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
     @Override
     public void beginTransaction(ConnectionContext context, TransactionId xid) throws Exception {
         super.beginTransaction(context, xid);
-        if (xid.isXATransaction()) {
-            replicateBeginTransaction(context, xid);
-        }
+        replicaEventReplicator.replicateBeginTransaction(context, xid);
     }
 
     @Override
     public void forgetTransaction(ConnectionContext context, TransactionId transactionId) throws Exception {
         super.forgetTransaction(context, transactionId);
-        if (transactionId.isXATransaction()) {
-            replicateForgetTransaction(context, transactionId);
-        }
-    }
-
-    private boolean needToReplicateAck(ConnectionContext connectionContext, MessageAck ack, PrefetchSubscription subscription) {
-        if (isReplicaContext(connectionContext)) {
-            return false;
-        }
-        if (ReplicaSupport.isReplicationDestination(ack.getDestination())) {
-            return false;
-        }
-        if (ack.getDestination().isTemporary()) {
-            return false;
-        }
-        if (subscription instanceof QueueBrowserSubscription && !connectionContext.isNetworkConnection()) {
-            return false;
-        }
-
-        return true;
+        replicaEventReplicator.replicateForgetTransaction(context, transactionId);
     }
 
     @Override
     public Destination addDestination(ConnectionContext context, ActiveMQDestination destination, boolean createIfTemporary)
             throws Exception {
         Destination newDestination = super.addDestination(context, destination, createIfTemporary);
-        if (shouldReplicateDestination(destination)) {
-            replicateDestinationCreation(context, destination);
-        }
+        replicaEventReplicator.replicateDestinationCreation(context, destination);
         return newDestination;
     }
 
     @Override
     public void removeDestination(ConnectionContext context, ActiveMQDestination destination, long timeout) throws Exception {
         super.removeDestination(context, destination, timeout);
-        replicateDestinationRemoval(context, destination);
+        replicaEventReplicator.replicateDestinationRemoval(context, destination);
     }
 
     @Override
@@ -663,7 +338,7 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
             return;
         }
 
-        if (!needToReplicateAck(connectionContext, ack, subscription)) {
+        if (!replicaEventReplicator.needToReplicateAck(connectionContext, ack, subscription)) {
             super.acknowledge(consumerExchange, ack);
             return;
         }
@@ -691,7 +366,7 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
 
         try {
             super.acknowledge(consumerExchange, ack);
-            replicateAck(connectionContext, ack, subscription, transactionId, messageIdsToAck);
+            replicaEventReplicator.replicateAck(connectionContext, ack, subscription, transactionId, messageIdsToAck);
             if (isInternalTransaction) {
                 super.commitTransaction(connectionContext, transactionId, true);
             }
@@ -715,83 +390,20 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
                 .collect(Collectors.toList());
     }
 
-    private void replicateAck(ConnectionContext connectionContext, MessageAck ack, PrefetchSubscription subscription,
-            TransactionId transactionId, List<String> messageIdsToAck) throws Exception {
-        try {
-            TransactionId originalTransactionId = ack.getTransactionId();
-            ActiveMQDestination destination = ack.getDestination();
-            ReplicaEvent event = new ReplicaEvent()
-                    .setEventType(ReplicaEventType.MESSAGE_ACK)
-                    .setEventData(eventSerializer.serializeReplicationData(ack))
-                    .setTransactionId(transactionId)
-                    .setReplicationProperty(ReplicaSupport.MESSAGE_IDS_PROPERTY, messageIdsToAck)
-                    .setReplicationProperty(ReplicaSupport.IS_ORIGINAL_MESSAGE_SENT_TO_QUEUE_PROPERTY,
-                            destination.isQueue())
-                    .setReplicationProperty(ReplicaSupport.ORIGINAL_MESSAGE_DESTINATION_PROPERTY,
-                            destination.toString())
-                    .setReplicationProperty(ReplicaSupport.IS_ORIGINAL_MESSAGE_IN_XA_TRANSACTION_PROPERTY,
-                            originalTransactionId != null && originalTransactionId.isXATransaction());
-            if (destination.isTopic() && subscription instanceof DurableTopicSubscription) {
-                event.setReplicationProperty(ReplicaSupport.CLIENT_ID_PROPERTY, connectionContext.getClientId());
-                event.setReplicationProperty(ReplicaSupport.SUBSCRIPTION_NAME_PROPERTY,
-                        ((DurableTopicSubscription) subscription).getSubscriptionKey().getSubscriptionName());
-                event.setVersion(3);
-            }
-
-            enqueueReplicaEvent(connectionContext, event);
-        } catch (Exception e) {
-            logger.error("Failed to replicate ack messages [{} <-> {}] for consumer {}",
-                    ack.getFirstMessageId(),
-                    ack.getLastMessageId(),
-                    ack.getConsumerId(), e);
-            throw e;
-        }
-    }
-
     @Override
     public void queuePurged(ConnectionContext context, ActiveMQDestination destination) {
         super.queuePurged(context, destination);
         if(!ReplicaSupport.isReplicationDestination(destination)) {
-            replicateQueuePurged(context, destination);
+            replicaEventReplicator.replicateQueuePurged(context, destination);
         } else {
             logger.error("Replication queue was purged {}", destination.getPhysicalName());
-        }
-    }
-
-    private void replicateQueuePurged(ConnectionContext connectionContext, ActiveMQDestination destination) {
-        try {
-            enqueueReplicaEvent(
-                    connectionContext,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.QUEUE_PURGED)
-                            .setEventData(eventSerializer.serializeReplicationData(destination))
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate queue purge {}", destination, e);
         }
     }
 
     @Override
     public void messageExpired(ConnectionContext context, MessageReference message, Subscription subscription) {
         super.messageExpired(context, message, subscription);
-        replicateMessageExpired(context, message);
-    }
-
-    private void replicateMessageExpired(ConnectionContext context, MessageReference reference) {
-        Message message = reference.getMessage();
-        if (!isReplicatedDestination(message.getDestination())) {
-            return;
-        }
-        try {
-            enqueueReplicaEvent(
-                    context,
-                    new ReplicaEvent()
-                            .setEventType(ReplicaEventType.MESSAGE_EXPIRED)
-                            .setEventData(eventSerializer.serializeReplicationData(message))
-            );
-        } catch (Exception e) {
-            logger.error("Failed to replicate discard of {}", reference.getMessageId(), e);
-        }
+        replicaEventReplicator.replicateMessageExpired(context, message);
     }
 
     @Override
@@ -802,19 +414,5 @@ public class ReplicaSourceBroker extends MutativeRoleBroker {
         }
 
         return super.sendToDeadLetterQueue(context, messageReference, subscription, poisonCause);
-    }
-
-    private void enqueueReplicaEvent(ConnectionContext connectionContext, ReplicaEvent event) throws Exception {
-        if (isReplicaContext(connectionContext)) {
-            return;
-        }
-        if (!initialized.get()) {
-            return;
-        }
-        replicationMessageProducer.enqueueIntermediateReplicaEvent(connectionContext, event);
-    }
-
-    private boolean isReplicaContext(ConnectionContext initialContext) {
-        return initialContext != null && ReplicaSupport.REPLICATION_PLUGIN_USER_NAME.equals(initialContext.getUserName());
     }
 }
